Replace absolute imports by relative ones
