+++ optimum/onnxruntime/__init__.py
@@ -11,44 +11,10 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from enum import Enum
 from typing import TYPE_CHECKING
 
 from transformers.utils import _LazyModule
 
-
-class ORTQuantizableOperator(Enum):
-    # Common ops
-    Gather = "Gather"
-    Transpose = "Transpose"
-    EmbedLayerNormalizationQuant = "EmbedLayerNormalization"
-
-    # QLinearOps
-    Conv = "Conv"
-    MatMul = "MatMul"
-    Add = "Add"
-    Mul = "Mul"
-    Relu = "Relu"
-    Clip = "Clip"
-    LeakyRelu = "LeakyRelu"
-    Sigmoid = "Sigmoid"
-    MaxPool = "MaxPool"
-    GlobalAveragePool = "GlobalAveragePool"
-    Split = "Split"
-    Pad = "Pad"
-    Reshape = "Reshape"
-    Squeeze = "Squeeze"
-    Unsqueeze = "Unsqueeze"
-    Resize = "Resize"
-    AveragePool = "AveragePool"
-    Concat = "Concat"
-
-
-AUTO_MINIMUM_SUPPORTED_ONNX_OPSET = None
-
-# This value is used to indicate ORT which axis it should use to quantize an operator "per-channel"
-ORT_DEFAULT_CHANNEL_FOR_OPERATORS = {"MatMul": 1}
-ORT_FULLY_CONNECTED_OPERATORS = ["MatMul", "Add"]
 
 _import_structure = {
     "configuration": ["ORTConfig"],
@@ -63,6 +29,12 @@ _import_structure = {
     ],
     "modeling_seq2seq": ["ORTModelForSeq2SeqLM"],
     "optimization": ["ORTOptimizer"],
+    "quantizable_operators": [
+        "ORTQuantizableOperator",
+        "AUTO_MINIMUM_SUPPORTED_ONNX_OPSET",
+        "ORT_DEFAULT_CHANNEL_FOR_OPERATORS",
+        "ORT_FULLY_CONNECTED_OPERATORS",
+    ],
     "quantization": ["ORTQuantizer"],
     "trainer": ["ORTTrainer"],
     "trainer_seq2seq": ["ORTSeq2SeqTrainer"],
@@ -84,6 +56,12 @@ if TYPE_CHECKING:
     )
     from .modeling_seq2seq import ORTModelForSeq2SeqLM
     from .optimization import ORTOptimizer
+    from .quantizable_operators import (
+        AUTO_MINIMUM_SUPPORTED_ONNX_OPSET,
+        ORT_DEFAULT_CHANNEL_FOR_OPERATORS,
+        ORT_FULLY_CONNECTED_OPERATORS,
+        ORTQuantizableOperator,
+    )
     from .quantization import ORTQuantizer
     from .trainer import ORTTrainer
     from .trainer_seq2seq import ORTSeq2SeqTrainer
