+++ optimum/onnxruntime/optimization.py
@@ -25,8 +25,9 @@ from onnx import load_model
 from onnxruntime.transformers.fusion_options import FusionOptions
 from onnxruntime.transformers.onnx_model_bert import BertOnnxModel
 from onnxruntime.transformers.optimizer import get_fusion_statistics, optimize_model
-from optimum.onnxruntime.configuration import OptimizationConfig
-from optimum.onnxruntime.utils import ORTConfigManager
+
+from .configuration import OptimizationConfig
+from .utils import ORTConfigManager
 
 
 LOGGER = logging.getLogger(__name__)
